Split path validation into a separate library
