+++ pkg/api/validation/path/name.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package validation
+package path
 
 import (
 	"fmt"
